fix type errors
