+++ src/components/App.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState, useRef } from "react";
 import ImportButton from "./ImportButton";
 import ExportButton from "./ExportButton";
 import ImageCanvas from "./ImageCanvas";
@@ -6,37 +6,26 @@ import "../styles/app.scss";
 import { image2hex } from "../lib/converter";
 import { saveAs } from "file-saver";
 
-class App extends React.Component {
-  // imageFile: File | null
-  canvasRef: React.RefObject<HTMLCanvasElement>;
-  imageRef: React.RefObject<HTMLImageElement>;
-  state: {
-    imageFile: File | null;
-  };
+///////////// Type Definitions:
+type ImageFile = File | null;
 
-  constructor(props: {}) {
-    super(props);
-    this.canvasRef = React.createRef();
-    this.imageRef = React.createRef();
-    this.state = {
-      imageFile: null
-    };
+
+function App(): JSX.Element {
+  const [imageFile, setImageFile] = useState<ImageFile>(null);
+  const canvasRef = useRef<HTMLCanvasElement>(null);
+  const imageRef = useRef<HTMLImageElement>(null);
+
+  const handleImageChange = async (imageFile: ImageFile): Promise<void> => {
+    setImageFile(imageFile);
   }
 
-  handleImageChange = async (imageFile: any) => {
-    this.setState({
-      imageFile: imageFile
-    });
-    // console.log(this.state.imageFile);
-  };
-
-  handleImageExport = () => {
+  const handleImageExport = (): void => {
     const alertMsg = () => alert("Please import an image first!");
-    if (!this.state.imageFile) {
+    if (!imageFile) {
       alertMsg();
     } else {
-      let canvas = this.canvasRef.current;
-      let image = this.imageRef.current;
+      let canvas = canvasRef.current;
+      let image = imageRef.current;
 
       if (!canvas || !image) {
         alertMsg();
@@ -51,8 +40,8 @@ class App extends React.Component {
       }
 
       let imageData = context.getImageData(0, 0, image.width, image.height);
-      let hexData = image2hex(imageData.data, this.state.imageFile.name);
-      let fileName = this.state.imageFile.name;
+      let hexData = image2hex(imageData.data, imageFile.name);
+      let fileName = imageFile.name;
       let fileType = ".c";
       let fullFileName =
         fileName.slice(0, fileName.lastIndexOf(".")) + fileType;
@@ -61,43 +50,39 @@ class App extends React.Component {
       console.log(hexData);
       saveAs(blob, fullFileName);
     }
-  };
+  }
 
-  render() {
-    return (
-      <div className="app-container">
-        <div className="navbar">
-          <span className="title">VOCC</span>
-          <span className="subtitle">
-            Game Boy Advance Image Editor and Converter
-          </span>
-          <ImportButton
-            onImageChange={(imageFile: File) =>
-              this.handleImageChange(imageFile)
-            }
-          />
-          <ExportButton startImageExport={() => this.handleImageExport()} />
+  return (
+    <div className="app-container">
+      <div className="navbar">
+        <span className="title">VOCC</span>
+        <span className="subtitle">
+          Game Boy Advance Image Editor and Converter
+        </span>
+        <ImportButton
+          handleImageChange={handleImageChange}
+        />
+        <ExportButton startImageExport={handleImageExport} />
+      </div>
+      <div className="workspace-container">
+        <div className="left-panel">
+          <div className="panel-label">Tools</div>
         </div>
-        <div className="workspace-container">
-          <div className="left-panel">
-            <div className="panel-label">Tools</div>
-          </div>
-          <div className="image-container">
-            {this.state.imageFile ? (
-              <ImageCanvas
-                imageFile={this.state.imageFile}
-                canvasRef={this.canvasRef}
-                imageRef={this.imageRef}
-              />
-            ) : null}
-          </div>
-          <div className="right-panel">
-            <div className="panel-label">Color Palette</div>
-          </div>
+        <div className="image-container">
+          {imageFile ? (
+            <ImageCanvas
+              imageFile={imageFile}
+              canvasRef={canvasRef}
+              imageRef={imageRef}
+            />
+          ) : null}
+        </div>
+        <div className="right-panel">
+          <div className="panel-label">Color Palette</div>
         </div>
       </div>
-    );
-  }
+    </div>
+  );
 }
 
 export default App;
